Move TypeCollector to the ir::traversal module
