+++ src/ir/context.rs
@@ -5,9 +5,8 @@ use super::int::IntKind;
 use super::item::{Item, ItemSet, ItemCanonicalPath};
 use super::item_kind::ItemKind;
 use super::module::{Module, ModuleKind};
-use super::traversal::ItemTraversal;
+use super::traversal::{ItemTraversal, TypeCollector};
 use super::ty::{FloatKind, TemplateDeclaration, Type, TypeKind};
-use super::type_collector::TypeCollector;
 use BindgenOptions;
 use cexpr;
 use chooser::TypeChooser;
