+++ src/ir/mod.rs
@@ -16,6 +16,5 @@ pub mod layout;
 pub mod module;
 pub mod traversal;
 pub mod ty;
-pub mod type_collector;
 pub mod var;
 pub mod objc;
